score_system/requirements.txt
aniso8601==7.0.0
atomicwrites==1.3.0
attrs==19.1.0
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.0
importlib-metadata==0.18
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
MarkupSafe==1.1.1
more-itertools==7.2.0
mysqlclient==1.4.2.post1
packaging==19.0
pluggy==0.12.0
py==1.8.0
PyMySQL==0.9.3
pyparsing==2.4.0
pyrsistent==0.15.3
pytest==5.0.1
pytz==2019.1
PyYAML==5.1.1
redis==3.2.1
six==1.12.0
SQLAlchemy==1.3.6
uWSGI==2.0.18
wcwidth==0.1.7
Werkzeug==0.15.4
zipp==0.5.2
